Create utility file for reusable functions, move formatTime into utility
