+++ src/components/DetailedRoute/DetailedRoute.tsx
@@ -6,6 +6,8 @@ import MapDisplay from '../Map/Map';
 import Days from '../Days/Days';
 import { geocodeByAddress, getLatLng } from 'react-places-autocomplete';
 import ProfilePhoto from '../ProfilePhoto/ProfilePhoto';
+import { formatTime } from '../../utilities/utilities';
+import { format } from 'util';
 
 interface DetailedRouteProps {
   userId: string
@@ -75,7 +77,9 @@ const DetailedRoute: FC<DetailedRouteProps> = ({ userId, currentUser }) => {
         </article>
         <article className='user-details__route-details'>
           <h3>Time: </h3>
-          <p data-cy='user-time'>{matchedUser?.user_rides[0].departure_time}</p>
+          <p data-cy='user-time'>
+            {matchedUser?.user_rides[0].departure_time && formatTime(matchedUser.user_rides[0].departure_time)}
+          </p>
           <h3>Days: </h3>
           {matchedUser && <Days matchedDays={matchedUser.ride_days} />}
         </article>
